Remove shared memory from GPU decompress
Now, each thread works autonomously, meaning it reads both the exponent
and the relevant compressed blocks without syncronizing with the others.
It is relied upon the cache to read the exponent value only once and
distribute it to the other threads.

Note: Does not work with non power-of-two bits_per_value because the
      compression was not adapted (same is true for the previous
      commit).